requirements.txt
# EduBot — Dependências
# Framework
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9

# Banco de dados
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
alembic==1.13.0

# HTTP client (para Claude API)
httpx==0.27.0

# Validação
pydantic==2.9.0
pydantic-settings==2.5.0

# Scheduler (notificações)
celery==5.4.0
redis==5.1.0

# Utilitários
python-dotenv==1.0.0
